+++ tsconfig.json
@@ -33,5 +33,9 @@
   "exclude": ["node_modules"],
   "ts-node": {
     "esm": true,
+    "experimentalSpecifierResolution": "node",
+    "compilerOptions": {
+      "module": "commonjs",
+    },
   },
 }
